+++ make-R.R
@@ -1,0 +1,8 @@
+rm(list=ls())
+gc()
+rstudioapi::restartSession()
+
+devtools::document()
+devtools::check()
+.path <- devtools::build()
+rcmdcheck::rcmdcheck(.path)