+++ tests/testthat.R
@@ -1,4 +1,7 @@
+library(dplyr)
 library(testthat)
 library(mrgda)
 
 test_check("mrgda")
+
+